CADRE SURVEY  - PRASAD

+++ partyanalyst-service/trunk/src/main/java/com/itgrids/partyanalyst/dao/IBoothPublicationVoterDAO.java
@@ -684,5 +684,5 @@ public interface IBoothPublicationVoterDAO extends
 	  
 	  public List<Object[]> getTotalVotersAndBoothTypeForConstituencyByBoothWise(Long constituencyId);
 	  public List<Object[]> getConstyPublicationIdByVoterId1(String voterCardNo);
-	  public Integer getTotalVoterByBooths(List<Long> boothIds);
+	  public Long getTotalVoterByBooths(List<Long> boothIds);
 }
